+++ sex_api/tests/test_searching.py
@@ -1,12 +1,20 @@
+import pytest
 from ..api import Client, Mode, Relevance
 
-client = Client()
-searching = client.search("Mia Khalifa")
+
+@pytest.fixture
+def client():
+    return Client()
 
 
-def test_searching():
-    for idx, pin in enumerate(searching):
+@pytest.mark.asyncio
+async def test_searching(client):
+    searching = client.search("Mia Khalifa")
+
+    idx = 0
+    async for pin in searching:
         assert isinstance(pin.name, str) and len(pin.name) > 3
 
         if idx == 5:
             break
+        idx += 1