+++ sex_api/tests/test_user.py
@@ -1,10 +1,15 @@
+import pytest
 from ..api import Client
 
-client = Client()
-user = client.get_user("https://www.sex.com/user/twitchxxx")
+
+@pytest.fixture
+def client():
+    return Client()
 
 
-def test_user_attributes():
+@pytest.mark.asyncio
+async def test_user_attributes(client):
+    user = await client.get_user("https://www.sex.com/user/twitchxxx")
     assert isinstance(user.count_pins, str)
     assert isinstance(user.count_repins, str)
     assert isinstance(user.count_following, str)
@@ -14,25 +19,33 @@ def test_user_attributes():
     assert isinstance(user.description, str) and len(user.description) > 3
 
 
-def test_user_pins():
+@pytest.mark.asyncio
+async def test_user_pins(client):
+    user = await client.get_user("https://www.sex.com/user/twitchxxx")
     pins = user.get_pins()
     liked = user.get_liked_pins()
     repins = user.get_repins()
 
-    for idx, pin in enumerate(pins):
+    idx = 0
+    async for pin in pins:
         assert isinstance(pin.name, str) and len(pin.name) > 1
 
         if idx == 5:
             break
+        idx += 1
 
-    for idx, pin in enumerate(liked):
+    idx = 0
+    async for pin in liked:
         assert isinstance(pin.name, str) and len(pin.name) > 1
 
         if idx == 5:
             break
+        idx += 1
 
-    for idx, pin in enumerate(repins):
+    idx = 0
+    async for pin in repins:
         assert isinstance(pin.name, str) and len(pin.name) > 1
 
         if idx == 5:
             break
+        idx += 1